Rename originWhitelist.js
